Remove PHP's open_basedir and custom temp dir on death.
The goal of open_basedir is to provide some kind of security (limiting
PHP file access to a certain directory), but...

- PHP runs as the user, so it's already limited to anything the user
  can do via their normal shell

- we allow system/exec/passthru/etc. so you can still do stuff like
  passthru("ls /home/") from PHP

- we allow arbitrary CGI scripts

- we allow user php.ini, so anybody could just turn it off

Given all that, open_basedir provides no real security for us, and just
creates problems (e.g. requiring the temporary directory to be inside
/services/, which is exported by NFS).

Removing the open_basedir restriction results in no real loss of
security, but creates less pain for us.

This resolves RT#1467 and is a permanent fix for the issue reported
in RT#1448.