Disable pam_delay on desktops to fix lightdm freezing
Temporary fix to prevent lightdm from freezing for ~120 seconds after an
invalid password is entered. Disabling pam_delay avoids this.

This is a temporary fix (hence it being placed in a manifest which is
only tangentially related to the problem) which should be removed after
the actual source of the problem is found and fixed.