Add accountsservice to desktops to work around lightdm bug
This appears to fix the bug where the lightdm greeter hangs for several
minutes after an incorrect password is given for an account. I don't
have a good reason for why this works (and stumbled upon it by
accident when testing gdm3, which brought it in as a dependency), but it
seems reliable after testing on a few of the lab machines.

accountsservice _is_ suggested by the lightdm package.

Cross your fingers...