Update PXE/netboot gen script for wheezy
Changing the `dist` variable was the only thing necessary for the script
to download and prepare a wheezy netboot (instead of squeeze). It's
actually a really nice script.

It now also adds an "OCF Automated Install" option which is selected by
default, and automatically chosen after a 10 second delay. This means
that simply PXE booting will result in a fully-installed system without
touching the keyboard at all, which is useful for spinning up VMs.

The preseed file (~staff/public_html/preseed.cfg) is fed to the
installer by DHCP.