Move environmentpath from [master] to [main]
For some reason, setting it in [master] works beautifully when you run
the master via command-line "puppet master", but doesn't work at all
when using Passenger.

Somewhat related to the bug report below, although they only addressed
making `puppet print` show the right value, and not that the value was
ignored by Passenger, which makes me think I'm missing something and
[master] isn't read by Passenger-backed puppetmaster (even though from
reading the docs it seems like it should be).

https://tickets.puppetlabs.com/browse/PUP-2213