Backport kernel on desktops for Haswell graphics
The integrated graphics on the new Haswell machines in the lab are,
unfortunately, extremely unstable with the 3.2 kernel currently in use
by stable.

Logging in works (although very slow), but the entire system crashes [1]
when logging out or when switching to a tty.

Installing a backported kernel fixes the problems entirely (both the
crashing and extremely poor performance).

[1] I didn't investigate what is actually happening since performance is
unbearably slow (to the point that fixing the crashing would not make
the desktop usable), but the screen goes dark, all network connectivity
is lost, and about 30 seconds later it restarts itself.