Don't keep puppet/puppetmaster updated via puppet
Having puppet update itself was creating problems with dependencies on
the puppetmaster (new puppet would be installed which conflicted with
the old puppetmaster, so puppetmaster was removed).

Instead of keeping these packages up-to-date with puppet, we will update
them the same way we update all other packages (using apt-dater), which
should be much more reliable.