Use backported xserver-xorg-video-intel on desktops
Backporting the kernel seems to fix all/most of the stability and
performance problems experienced with the new Haswell desktops, but
there are still a couple problems:

  - multiple displays aren't recognized (seen as one display with
    mirroring on the other)

  - on resume from suspend, screen is black and only repainted as-needed
    (e.g. when moving the mouse)

Installing the "latest" [1] intel drivers fixes these issues.

There is no official backport, so I backported the current version from
jessie, which seems to be working quite well. No dependencies had to be
backported, so this seems like a reasonable compromise while we wait for
jessie.

[1] latest available in testing; not even close to intel's latest
    release, of course