New polkit configuration
No longer modify the polkit action files directly. Instead, properly configure
the polkit local authority to deny all actions except mounting and ejecting
external drives.

This configuration is compatible with polkit <= 0.105, as currently included
in wheezy and jessie, but not with 0.106 and above which use JavaScript rules.
It will need to be revised in the future.

Other changes:
  - Add support for udisks2 which is used in jessie.
  - Configure ocfroot group for polkit administrator authentication.