Mirror debian-cd on fallingrocks
We still have some spare disk space, so why not?

Ideally we would do the whole jigdo-mirror setup, but the online docs
are terrible and outdated (and don't appear to be what the official
mirrors are using) and I don't think it's worth it to avoid a couple
hundred GB of download every few months. If anyone disagrees, feel free
to explore.