desktop: use only 1 workspace (xfce)
(cherry-picked from jessie branch)

Users (and me) get confused when scrolling on the desktop switches to
another workspace. We could show a workspace applet and disable
desktop scrolling, but I think it will still be confusing.

If anybody actually uses workspaces we could do that, but not sure that
anybody does?

I spent two hours writing this using augeaus and then realized we're
syncing the entire directory, so that won't even work. For the record,

    augeas { 'xfwm4':
      lens    => 'Xml.lns',
      incl    => '/etc/xdg/xfce4/xfconf/xfce-perchannel-xml/xfwm4.xml',
      changes => [
        "set channel[#attribute/name='xfwm4']/property[#attribute/name='general']/property[#attribute/name='workspace_count']/#attribute/value 1"
      ];
    }

We could not sync the entire directory (probably better), but this file (and
xfce4-panel.xml) don't even get created by default, so we'd have to recreate
the entire file with augeaus (unless it just overrides default configs? who
knows...)