ocf_mirrors: add basic health checks
These may not be perfect, but I think they will give us some idea. I'm
not entirely sure if my understanding of "Date" is correct, so we might
need to change that later if we start getting spammed with alerts
incorrectly.

(I'm actually pretty sure that my understanding is wrong and so we might
get false positives, but not sure what else we can easily check to see
if we are receiving updates.)