ocf_www: switch from mod_fcgid to mod_fastcgi
Using mod_fastcgi allows us to use tha "autoRestart" option to
automatically restart a FastCGI process when the mtime of the script is
updated (i.e. when touching the script).

Note that this requires a change to the vhost config files, which
currently use FcgidWrapper to define a specific wrapper for php scripts.