ocf_www: add php5 fcgi wrapper source
We use a wrapper for virtual hosts to provide PHP via FastCGI (instead
of suPHP). This is significantly faster than suPHP, and not much more
complicated than not doing it.

This also allows PHP-using vhosts to benefit from APC. The scripts which
build vhost definitions are in ~staff/vhost and currently not in
revision control (they should probably be moved to utils).