ocf_www: enable actions module
We use the actions module to use a specific FastCGI wrapper for .php
files for virtual hosts. Virtual hosts are special because they each get
their own FastCGI PHP wrapper; we want to use this wrapper so that
vhosts get improved PHP performance (can benefit from APC, etc.).

This change is necessary because of the switch to mod_fastcgi, which,
unlike mod_fcgid, doesn't support defining a wrapper script for only
files of a certain suffix. So, instead, we add to each vhost:

    AddHandler php5-fcgi .php
    Action php5-fcgi /php5-fcgi
    Alias /php5-fcgi /opt/suexec/$login/php5-fcgi-wrapper

...which achieves the same thing.