ocf_dhcp: switch yearbook ip
yearbook.ocf.b.e is 169.229.10.216, which is an address we reserved for
them. I'm not sure why I gave them 130 instead -- probably a mistake.

(For context, we provide internet to B&G; they have a router which does
NAT in their space, which is patched into our server room. This is
probably only temporary and won't continue in the new space (not that
it's really much of a burden on us, though).)