ocf_apphost: add daemontools for process supervision
This is the best way I can find for users to supervise their processes.
They can just put a line in their crontab:
  @reboot svscan ~/apps
...and now when biohazard is restarted, their app will be started (and
later restarted as needed). It also makes it a lot easier to signal the
process (svc -h ~/apps/foo for SIGHUP, etc.).

One thing that isn't quite ideal is that you can't easily attach to the
running process to see what it's spitting out on stdout/stderr (like you
could if you ran it in screen/tmux). But competent people will be able
to figure out how to get to their logs, and everyone else shouldn't be
on the apphost anyway.

(I'm sorry Timmy)