common/auth: enable root password login
By default, jessie disables root login via SSH except via SSH keys.
It's actually important that we be able to SSH in to a server as root in
case Kerberos goes down [1], so we explicitely turn it back on.

In the future, maybe we should find a good way to use SSH keys (rather
than a shared root password) for the root user.

[1] Technically if LDAP goes down, we should still be able to log in due
to nscd, but it might take a long time. Kerberos is necessary for
password/GSSAPI auth, though.