desktop/xsession: add back penguin to lightdm
Unfortunately it's somewhat hacked-in because there's a bug in current
lightdm-gtk-greeter that prevents us from using the normal
default-user-image property to specify an image:

https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=767576

This is already fixed upstream, but the new version wouldn't have time
to make it into jessie before the freeze even if it was submitted today.
So, for now, we just use this relatively hacky solution.