common/apt: use hack to work around apt pinning bug
The puppetlabs apt module sets up apt priority pinning for backports
with a file like this:

    root@raptors:/etc/apt/preferences.d# cat backports.pref
    Explanation: apt: backports
    Package: *
    Pin: origin mirrors
    Pin-Priority: 200

See the problem? It uses hostname-based pinning, which won't work if we
use the same hostname for the backports archive as the main archive
(like in our case -- and, I suspect, in most peoples', now that
backports are merged into the main Debian package archives).

All packages from the mirrors hostname get the same priority, so it
tries to upgrade every package to its backport (if it has one), since
the version numbers are greater.

This is a terrible hack which switches the hostname to something
different so that the hostname-based pinning actually works. I'll work
on filing a bug report (and possiblty submitting a patch) to the
puppetlabs-apt module to fix this broken behavior.