desktop: don't overwrite pam common-auth and login
In b682f5, a naive me thought to remove pam_authdelay by removing the
line from two pam settings files and puppeting the entire giant files.

As Dara points out, this is what actually broke pam-auth-update (in
addition to being messy). It was only intended to work around a
persistent lightdm freezing bug, which is (hopefully) no longer an
issue. So, we can remove these files.

(If we decide to disable pam_authdelay again later -- which isn't really
a bad idea on the desktops -- we should do it not like this.)