ocf_dhcpd: remove separate pool for kvm
The KVM pool originated when makevm assumed the DHCP server assigned
VMs using DHCP a single IP address. makevm connected to the server on
this IP, assuming it was the new VM.

The pool of more than one IP was a somewhat unreliable hack to allow
multiple VMs to be spawned simultaneously. makevm no longer cares about
a small pool, and instead uses the ARP-advertised IP address of the VM
(matched with its MAC address), which is far more reliable.

So, we no longer need a special pool for KVM, and can just move those
addresses into the main pool.