desktop/grub: fix unrestricted boot on wheezy
c1b21cb mistakingly added --unrestricted to all GRUB boot options, with
the intention to not require the GRUB username/password on boot.

This is the correct approach on jessie, although it's not necessary on
wheezy. Unfortunately, the version of grub2 on wheezy doesn't even
recognize the option and fails to boot, even after entering the grub
username/password.

So, we ensure that `--restricted` is not present in grub configs to
clean up the previous mistake.