ocf_ssh: mount /tmp as tmpfs, add class dependencies
For users who want to work with lots of data (> 2GB), it might be useful
for them to have a place to do that outside of their homse, even if it's
not persisted. /tmp is a reasonable place for that.

Also moves the class dependencies which were previously only in LDAP to
the manifest.