ocf/repackage: improve handling of backports
This should be a more reasonable approach to handling backports and
packages without recommends (probably better than `dpkg -l | grep ^ii`,
anyway).

The manifest isn't super pretty because Puppet's limitations (can't
modify/append to variables in the same scope) make writing this the
obvious way hard/impossible.

I'm open to suggestions on improving this, but I suspect this approach
is preferable to the old one.

(With that said, I was a big fan of the whole...
    $b = "-t ${dist}-backports ", $g = '| grep ~bpo'
    aptitude -y ${r}${b}install
magic)