extrapackages: remove php5
php5 is not, as you might reasonably suspect, a package which installs php5.

Unlike basically every other language package, php5 is ``a metapackage that
[...] guarantees that you have at least one [...] server-side versions of the
PHP5 interpreter installed.''

php5 depends on:
    libapache2-mod-php5 | libapache2-mod-php5filter | php5-cgi | php5-fpm
and cannot be satisfied by just php5-cli.

This pulled in libapache2-mod-php5 on all the desktops, plus tsunami,
supernova, and biohazard. This, of course, also pulled in all of Apache.