apt: add needrestart in auto mode on jessie
needrestart got pulled in on jessie, as it is recommended by
apt-dater-host. It lists services which need to be restarted after
package updates, and offers to restart them for you.

I think keeping it is sane, but we should put it to automatic mode,
since currently it prompts interactively before restarting services,
which is annoying when using apt-dater. Presumably when doing updates we
expect the services to restart, so I think changing it to automatic is
reasonable.

This resolves rt#3110.