groups.sh: lower FIRST_SYSTEM_UID to 100
ocfstaff has been renumbered to 1001, and since it was the only LDAP
group colliding with the normal system gid range, we can use the normal
one now (and stop breaking upgrades).

We should still eventually renumber the ocf group, but that's going to
be a lot more of a challenge.

See rt#3096 for context.