desktop: remove steam apt repo, get from nonfree
Dara points out that, even though the official downloadable Steam
installer adds the repo.steampowered.com apt repo, the installer is
actually packaged in Debian jessie nonfree. So we'll just use that.

It's a bit out-of-date already, but the binary basically just downloads
a few hundred MB of the latest Steam installer. So it's not really a
concern if the downloader is out-of-date.