desktop/steam: install from local package
There's a bug (Debian bug#772598) preventing preseeding acceptance of
license answers in the current steam package.

We use a custom-built package with my patch to avoid that. We can remove
a lot of the crappy exec code once we switch to using apt (either
because we finally set up our own apt repo, or because the bug gets
fixed in Debian).