ocf_mirrors: fix ubuntu health check to find latest dist
Apparently devel is not constantly updated, so we can't use it to check
if the mirror is being updated. Instead, we try to figure out the latest
dist automatically via some terrible scripting magic.

This would be so much easier if, like Debian, Ubuntu had symlinks to
dists with names like "stable" or "unstable".