ocf_www: move ocf and pma sites to /services
The primary purpose of this move is to get as much important info on NFS
(rather than on individual server disks) as possible. This makes things
like backups, bringing up more servers, etc. possible.

Note that I moved pma to NFS, but not hello -- that's because I intend
to make pma auto-deploy from GitHub in the very near future.