sudoers: add nopasswd option for sudoers access
This sets NOPASSWD in the sudoers entry for ocfstaff (and owners) for
staff VMs, the rationale being that people without ocfroot have root on
these, and so staff should not be entering their passwords on them.

I originally did this on the desktops as well, but later decided against
it since:
  - it's likely staff leave themselves logged in by accident on desktops
  - staff will need to enter password to log in anyway

This is actually not the reason I did this; it will be used on the
offsite server, since we won't have Kerberos to verify passwords when
firestorm is down.