ocf_www: make access logs world-readable again
With the transition to HTTPS-only, this is the only way users can see
their access logs (pending some other more techncial solution).

To address the inevitable response: I don't think anybody actually holds
an expectation that GET parameters are private just because they're
using HTTPS. If they are, they should stop.

I'd be ok with a more technical solution as long as it's relatively
simple (setuid binary that filters access logs maybe?) but this is the
practical one until someone does that work.