ocf_mail: add mail metadata logging
We used to log all mail that went through sandstorm for ~7 (?) days so
that it could be examined if we suspected spam. When I rebuilt the mail
server as anthrax, I didn't re-implement this.

Instead, I added full mail logging for "nomail"'d users, and no logging
for normal users.

This adds more logging without being overly intrusive. It logs time
sent, from/to, sender UID, and subject. Almost all of this information
was already logged by default to /var/log/mail.log, but in an
inconvenient-to-parse-or-search format.

This produces nice JSON logs:

{"date": "2015-02-04T17:40:06.049154",
 "to": [null, "ocfstats@ocf.berkeley.edu"],
 "from": ["Cron Daemon", "root@ocf.berkeley.edu"],
 "uid": "999",
 "subject": "Cron <ocfstats@dementors> /opt/stats/bin/toner > /opt/stats/toner"}

Eventually we can parse this every day or so to automatically flag users
sending too much mail.

The only additional info logged is subject and UID, which I think is a
reasonable compromise for privacy reasons, especially given that we used
to log *everything*.