ocf/webhook: add secrets directory
This also moves services into a `services` subdirectory, since we manage
those recursively.

Eventually it would be a nicer API if it accepted `secretpath` which
could be a Puppet resource like `puppet:///private/mysecret` but that
requires a bit more work (need some way to disambiguate multiple
webhooks; maybe by path? -- but that requires some escaping first, so
it's not totally trivial). With that change, it wouldn't be necessary to
first copy a secretfile manually, either.