ocflib: don't require requests, install pip-whl
There is a good history of this problem in rt#3268.

A summary is:

  * We currently install ocflib system-wide via pip, which results in
    dependencies like requests also being installed system-wide.

  * Updating requests breaks pip, since the version in jessie depends on
    a non-public API in requests, which was changed in newer versions.

  * pip was recently changed in jessie to load wheels for its
    dependencies onto the path (at the start), if they're installed. So,
    if we install the wheel packages, we don't have to worry about
    breaking pip.

Previously, we built a patched version of pip which didn't rely on the
internal requests API. But this is better because we don't need to build
any of our own packages.

A future goal is to install ocflib in such a way that it doesn't mess
with system python dependencies (rt#3403).