ocf_www: raise maximum FastCGI processes
When we hit the limit, new processes won't spawn, which results in PHP
on vhosts not loading, and other FastCGI (e.g. Flask) not loading.

A user reported today that their FastCGI app wasn't loading, confirmed
it in the logs. It looks like we were *just* hitting the limit (we're at
about ~110 now that I've raised it), so this should give us plenty of
breathing room.