ocf/repackage: specify dist to backport on
This is a little clunky, but allows us to be more selective with
backporting. Rather than be stuck using a backported package because
back in squeeze it had a bug (and the current staff having no idea why
the package is backported anyway), we can explicitly backport things
after upgrades as necessary.

Currently everything is backported only on wheezy. There are some
immediate candidiates for backports on jessie too:
* Iceweasel
* LibreOffice

...but haven't decided on these yet. We might want to add another syntax
for allowing backport_on to be an array (and also not sure if Puppet
allows this without the future parser?).