ocf/access.conf: allow remote root logins
Per discussion after the last time LDAP was down, it's really difficult
to log in as root remotely, to the point that if I didn't have a habit
of leaving myself SSH'd into a dozen servers at all times, we would have
been in trouble.

We have a very strong root password now (and it's enforced by Puppet and
consistent across servers), so there should be no danger in allowing
remote root logins.

It would probably be easier for targeted attackers to break some random
staffer's password. And random botnet attackers aren't going to guess
the root password anyway.