ocf_mail/spam: disable Mail::SpamAssassin::Plugin::AWL
This was disabled by default in new versions. Besides, it's broken in
jessie:
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=739738

We could patch if we desired, but since it's disabled anyway, I'd rather
not.

TODO: can we stop supplying our own v310.pre? The only different is that
we disable SpamCop, which seems reasonable, but perhaps we could disable
it in a different file rather than providing the entire v310.pre.