Update apt module to 2.0.1
This disables apt pinning in ocf::repackage because it now creates a
dependency cycle (pins are defined in stage "first"). We need to
reconsider how we handle backports anyway.

Pushing this out earlier than I'd like because of an apparent bug where
the puppetmaster and puppet agents aren't agreeing on the version of the
native (Ruby) apt_key type, which is causing tons of errors.