Add ocf_jenkins module
Right now the plan is to do most configuration from within Jenkins
itself. It looks like there are some options for configuring Jenkins via
YAML (or similar) which we could Puppet -- but I don't want to mess
around with that until we're more familiar with it, since we'd probably
just make a mess of it.

Jenkins is up and live:
https://jenkins.ocf.berkeley.edu/

You can log in with your OCF username and password (ocfroot has access
for now, will figure out a proper ACL at some point).