ocf_ssl: provide 2048-bit dh parameters
These are not secret, so there is no harm in commiting them here.

Generating these on-the-fly can take a *long* time (minutes), and is
potentially dangerous to do on the first Puppet run when entropy is low.

Sharing these between servers/services is safe. So we save ourselves the
trouble by generating them before-hand.

-- Chris "The NSA is Not Standing Behind Me" Kuehl