ocf_backups/mirror: add encrypted backup generation
This gets us a good deal of the way closer to having automatic off-site
backups. It works by creating an LVM snapshot of the backups mirror at
the current time and then dd-ing an encrypted version if it into the
scratch space at /opt/backups/scratch.

As of today, it takes 8hr 18m to make this backup, for a total of 823
GiB of compressed data (2.0 TiB uncompressed).

The remaining work is to upload the encrypted files to Google Nearline
or something similar (and delete the old backup(s) afterwards).