ocf_www/site/vhosts: add is_array check around vhosts
This is a quirk of Puppet 3 where stringify_facts is on by default. We
turn it on, but it won't be on at the start of the first run, and the
catalog compile fails (which is why it's not just sufficient to turn it
on in the first stage, like we do).