Replace xfce4 application menu with whisker menu
Also add shortcut to open the menu using the Super key

The whisker menu is much nicer to use than the default xfce4 menu, since
it can be searched, and has nice things shown like the user's name. It is
also very extensible, since custom actions can be added to it very easily.