ocf_stats: filter individual munin alerts
Munin has no way to disable alerts from a plugin on all nodes (for those which
have hard-coded warning levels), so it has to be done by script. This is more
fine-grained than silencing an entire plugin anyways. Also, contrary to the
documentation, the actual keyword to do so is 'none', not 'no'.