Install enforcer to enforce print quotas.
Currently, launch is blocked by the fact that cups/tea4cups adds a trailing
page to the document before it is sent to the printer to make sure it has an
even number of pages. Hopefully there's something in the printer driver or the
*2pdf config to fix this.

Also, PDF is now the backend filetype. Hopefully this turns out to be highly
compatible and flexible.