Switch to PostScript, properly count duplex jobs
Switching to PostScript was to fix a PDF file I found that got mangled by
pdf2pdf. Unfortunately page counting is much slower with enforcer-pc (due to
grep I think), but I don't know of a way around it and jobs are limited to 30
pages anyways.

Fixing duplex counting was trivial, but that same file also has the problem
that the printer asks for A4 paper to print the last two pages, and you have to
press OK on the machine to finish the job. This doesn't happen on pollution,
but then I have seen similar problems for some jobs in the past. Go figure.

It would probably be a good time to do a soft launch now.