Add "nossl" flag for vhosts that don't need SSL
We'll still get certs for them in case we change our minds (seems fine),
but this means we won't use those certs.

This came up because contrib.o.b.e had switched itself to HTTPS, but the
whole reason we created contrib was to be *not* HTTPS (for d-i's preseed
file). Good catch @jvperrin!