Fix dependency cycle with docker apt source
Ideally we will not use Puppet stages in the future, since
puppetlabs-apt specifically says it [does not
work](https://github.com/puppetlabs/puppetlabs-apt#limitations) with
them, which is likely causing these dependency cycle issues. I'm still
not sure how this fix actually fixes the dependency cycle, but it does,
so that's good.