Use per-expression defaults in modules
This contains the effect of defaults to the module in which they are
used, instead of escaping to other modules through dynamic scope.

See https://github.com/ocf/puppet/pull/58 for an example of why we might
want to switch away from using dynamic scope for defaults.

Site-wide manifests still use the original resource default statements
instead of the new kind, because they are meant to have a global effect,
so they do not need to be limited like the ones in modules need to.