Remove SSLCACertificatePath from vhosts
This was copied from the Puppet versions of the vhost files, but it
causes Apache to take a *long* time to start and often time out after 20
seconds.

strace reveals it was re-reading all of these files in this directory
for each vhost (hundreds of them).

We don't need this since we're not using client certificate
authentication.