Make Kali use the official rsync source
Since we're now whitelisted to sync from it! :D

Push mirroring also doesn't work without a login shell, so mirrors now
has one, however password-based login is disabled and only one command
can be run based on the key used to log in, so this shouldn't be a
security problem.