Use system users to prevent UID/GID collisions
Fixes rt#5991

I also had to change the UIDs and GIDs that collided, but that was just
a pretty simple find of all the files with the old UID/GID and reown
them as the proper user/group. This was complicated a little since ocf
and ocfstaff shouldn't be changed in most places, like home directories.
However, they were the groups that collided, but all that involved was
making the find query more localized to the directories used by the
system users.