Use mesos host IP in mesos-dns, not docker IP
This makes it possible to reach the container from other hosts that are
not the one actually running the container. Currently it just uses the
local IP of the container in mesos-dns:

$ dig rt-5w8ru-s0.marathon.mesos
;; ANSWER SECTION:
rt-5w8ru-s0.marathon.mesos. 1   IN      A       172.17.0.7

However, this local IP isn't very useful from outside the host running
that specific container, so we should use the IP of the host itself
instead.