ocf_mirrors: add systemd timer resource
The ocf::systemd::timer implementation creates a systemd
timer and a corresponding service. The service will
activate ones the timer has expired. Both the timer
and service are implemented using templates.

Furthermore, we create a more specific ocf_mirrors::timer
that utilizes the OnCalendar option -- the systemd analogue
to cron jobs.

Lastly, we ensure that if any timer resource is created,
a cron job under the same name is purged.