Simplify ftpsync health checks
Use one script that takes command line args for health checks as opposed to
creating a unique script for every mirror using a template. This way we take
advantage of the ocf_mirrors::monitoring defined type to handle url building
instead of doing it the exact same way in python.

In the spirit of the old style, each mirror has a health symlink pointing to
the actual script at /opt/mirrors/bin/health.