Don't run make for all new staff puppet envs
rt#7054 has more details, essentially we are running out of inodes on
our puppet master, and this will help to lessen the effects of that
problem since the third-party modules count for most of the inodes used
in each environment. Also most new staffers don't use puppet at all and
the README says to run `make` on first setup anyway.