Pin marathon-lb to a specific version (rt#6870)
This will make updating marathon-lb possible instead of the current
problem where upgrading marathon-lb is a very manual process, because
restarting ocf-lb doesn't work because it'll use the current version
downloaded already instead of fetching a new version. Pinning to a
specific version makes this both more stable and makes it possible to
upgrade when wanted.