remove mesos from crisis
docker does some weird things on ZFS that is causing problems in
booting the root-on-ZFS machines. For some reason it seems to be
interpreting the overlay FS as a persistent dataset on the root
filesystem, polluting it with thousands of entries, which
interferes with the root mounting process. I'm going to debug this
more later.