Add systemd service for restarting fcgi processes
This is definitely a bit janky, but I still think it's the best solution
I've come across and seems like the safest option, since it isn't
in the critical path of any requests. It's also faster and less
disruptive than just setting the Fcgid settings to have a low lifetime
for any running processes, because that would kill off any started PHP
processes too, which is undesirable.