Bootstrap kubernetes master and worker nodes
Using the puppetlabs-kubernetes module, add functionality to
boostrap a host as either a kubernetes master or a kubernetes
worker node. Note that adding additional master nodes will require
running puppetlabs kubetool on an OCF host to generate new etcd
keys and certs, and the generated file must be added to the
puppet private share in the kubernetes directory.

With this configuration we have successfully deployed the OCF
templates container.