Add session stickiness to HAProxy
This will do load balancing and ensure clients are sent to the same
server for applications where session affinity is important, like PMA.
The downside of using this method is that consistent hashing doesn't
provide a "perfect" hash. Some ingress workers may get more workers than
others. For our case this shouldn't be an issue, but we may want to
explore the option of using cookie based session affinity since all of
our services are HTTP. HAProxy can use application cookies or inject its
own. We would want to do this on a per-application basis.