Add ocf::systemuser resource
Setting system => true is something we have shown time and time again
that we are not able to remember, creating annoying ldap conflicts.
Having a custom resource for this will make it much more apparent. The
resource is simple. It proxies all the arguments passed to opts to the
system resource, but ensures that system is set to true and that the
user is in the group 'sys'.