kubernetes: add firewall rules to allow etcd and kube communication
Kubernetes requires some ports to be opened for masters to communicate
state. Furthermore, etcd needs ports 2379 for client communication and
etcd peers, which are all the kubernetes masters, communicate on 2380.

It is very unlikely an attacker could enter the network through one of
the masters. For that reason we are lenient and allow all ports for both
TCP and UDP.