kubernetes: fix etcdctl (#475)
etcdctl needs some environment variables to make with it work without
passing a million flags every time we use it. It needs to know the CA
and server certs so the client can authenticate it is talking to a valid
peer. etcd traffic is encrypted so it needs the server key to decrypt
information.

We also need to set the ETCD_ENDPOINT environment variable with the
_same_ endpoint that was passed to kubetool as part of the initial
cluster. This is because kubetool generates the certs for those values.
In our case we just used the hostname.