package: update brave
The Brave version on our desktop has been notifying us that the current
version is out-of-date. Firstly, the package was renamed from brave to
brave-browser, the upstream apt repository to add changed, and the gpg
key has been rolled. Brave recommends pulling the key from their server,
but instead I uploaded the key to our keyserver and fetch it from there.

The configuration for brave and the configuration options have also
completely changed. Instead of making modifications to
`~/.config/brave/session-store-1`, we now put configuration options in
`~/.config/BraveSoftware/Brave-Browser/Default/Preferences`. Since brave
is built on top of chromium, that means they inherit chromium's style of
asking to be the default browser, via a timestamp... meaning it will
always ask you forever until you set it to be the default browser. With
a small hack we circumvent this. In the chromium code they will consider
your config option 'unparseable' if you pass a negative value as the
timestamp, thus it never asks to be the default browser.

If interested, the code is here:
https://github.com/chromium/chromium/blob/2ca8c5037021c9d2ecc00b787d58a31ed8fc8bcb/chrome/browser/ui/startup/default_browser_prompt.cc#L90-L91. Get fucked, Google :)